+++ src/components/pages/career-details/new-career-form/career-form-fields.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
+import { ErrorCode, FileError } from 'react-dropzone';
 import { Trans, useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import { SectionHeader } from '../../../content/section-header/section-header';
 import { Checkbox } from '../../../forms/checkbox/checkbox';
+import { FileDropper } from '../../../forms/file-dropper/file-dropper';
 import { StyledErrorMessage } from '../../../forms/text-input/text-input';
 import { ContentBlockContainer } from '../../../layout/content-block-container';
 import { TextStyles } from '../../../typography';
@@ -10,6 +12,7 @@ import { Button } from '../../../ui/buttons/button';
 import { StyledLink } from '../../contact/contact-form';
 
 const PRIVACY_POLICY = 'https://satellytes.jobs.personio.de/privacy-policy';
+const MAX_SIZE = 20 * 1024 * 1024;
 
 const StyledButton = styled(Button)`
   margin-top: 48px;
@@ -65,6 +68,57 @@ export const CareerDetailsCheckbox = ({ control }) => {
       }
       control={control}
       rules={{ required: t<string>('career.error.approval') }}
+    />
+  );
+};
+
+export const CareerDetailsFileUpload = ({
+  register,
+  setValue,
+  setError,
+  clearErrors,
+  errors,
+}) => {
+  const { t } = useTranslation();
+
+  const fileCategories = [
+    {
+      value: 'cv',
+      label: t<string>('career.cv'),
+    },
+    {
+      value: 'cover-letter',
+      label: t<string>('career.cover-letter'),
+    },
+    {
+      value: 'other',
+      label: t<string>('career.other'),
+    },
+  ];
+
+  const fileValidator = (file: File): FileError | null => {
+    if (file.size > MAX_SIZE)
+      return {
+        message: t<string>('career.error.max-size'),
+        code: ErrorCode.FileTooLarge,
+      };
+
+    return null;
+  };
+
+  return (
+    <FileDropper
+      register={register}
+      setValue={setValue}
+      setError={setError}
+      clearErrors={clearErrors}
+      errors={errors}
+      name="documents"
+      fileCategories={fileCategories}
+      acceptedFileTypes={'.pdf'}
+      illustration="monitor_024"
+      maxFiles={3}
+      validator={fileValidator}
     />
   );
 };
